database schema changed, fixed bugs

+++ StrattonOakmont/StrattonOakmontModels/Users/Company.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Desc { get; set; }
-        public decimal Сapitalization { get; set; }
+        public long Сapitalization { get; set; }
         public string IMG { get; set; }
         public Security Security { get; set; } 
     }
